Remove types.go
