+++ src/app/backend/resource/thirdpartyresource/list.go
@@ -23,6 +23,19 @@ import (
 	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
 )
 
+// ThirdPartyResource is a third party resource template. Used in list view.
+type ThirdPartyResource struct {
+	ObjectMeta common.ObjectMeta `json:"objectMeta"`
+	TypeMeta   common.TypeMeta   `json:"typeMeta"`
+}
+
+// ThirdPartyResourceList contains a list of third party resource templates. Used in list view.
+type ThirdPartyResourceList struct {
+	ListMeta            common.ListMeta      `json:"listMeta"`
+	TypeMeta            common.TypeMeta      `json:"typeMeta"`
+	ThirdPartyResources []ThirdPartyResource `json:"thirdPartyResources"`
+}
+
 // GetThirdPartyResourceList returns a list of third party resource templates.
 func GetThirdPartyResourceList(client *internalclientset.Clientset,
 	dsQuery *dataselect.DataSelectQuery) (*ThirdPartyResourceList, error) {
