--- src/app/backend/resource/thirdpartyresource/types.go
@@ -1,40 +1,0 @@
-package thirdpartyresource
-
-import "k8s.io/kubernetes/pkg/apis/extensions"
-
-// Copyright 2015 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import "github.com/kubernetes/dashboard/src/app/backend/resource/common"
-
-// ThirdPartyResource is a third party resource template. Used in list view.
-type ThirdPartyResource struct {
-	ObjectMeta common.ObjectMeta `json:"objectMeta"`
-	TypeMeta   common.TypeMeta   `json:"typeMeta"`
-}
-
-// ThirdPartyResourceList contains a list of third party resource templates. Used in list view.
-type ThirdPartyResourceList struct {
-	ListMeta            common.ListMeta      `json:"listMeta"`
-	TypeMeta            common.TypeMeta      `json:"typeMeta"`
-	ThirdPartyResources []ThirdPartyResource `json:"thirdPartyResources"`
-}
-
-// ThirdPartyResourceDetail is a third party resource template. Used in detail view.
-type ThirdPartyResourceDetail struct {
-	ObjectMeta  common.ObjectMeta       `json:"objectMeta"`
-	TypeMeta    common.TypeMeta         `json:"typeMeta"`
-	Description string                  `json:"description"`
-	Versions    []extensions.APIVersion `json:"versions"`
-}
